refactor(server): rename campaignHousingService to kebab-case
Rename service file and test to follow kebab-case convention:
- campaignHousingService.ts → campaign-housing-service.ts
- Update all import paths in campaignController.ts and flip-sent-campaign-housings/task.ts
- Update test describe label to match new filename
